:sparkles::white_check_mark: add: getBalance

+++ src/main/kotlin/com/zbank/bankaccount/domain/model/account/AccountRepository.kt
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface AccountRepository : CrudRepository<Account, Int> {
+interface AccountRepository : CrudRepository<Account, Long> {
 
     fun existsByCpf(cpf: String): Boolean
 }
